switched to temp file instead of :memory: sqlite database
