+++ package.yaml
@@ -44,6 +44,7 @@ dependencies:
 - text
 - time
 - transformers
+- unix
 - unordered-containers
 
 library:
